requirements/app.txt
# Core dependencies
django==2.2.9
olefile==0.46
placebo==0.9.0
pycparser==2.19
psycopg2-binary==2.8.4
Jinja2~=2.11.1
pytz~=2019.3
python-dateutil==2.6.1
phonenumbers==8.11.2
django-storages==1.8
django-allauth==0.41
django-invitations==1.9.2
sendgrid-django~=2.0
django-jinja~=2.5.0
djangorestframework~=3.11.0
openpyxl==3.0.3
reportlab==3.5.34
toml==0.10.0
troposphere==2.5.3
typing==3.7.4.1
Pillow~=6.2
python-Levenshtein~=0.12
django-taggit==1.2.0
mixpanel~=4.5.0
django-extensions~=2.2.6
twilio==6.35.3
django-compressor==2.4.0
PyYAML==5.2
user-agents==2.0
ua-parser==0.8.0
django-easy-audit==1.2.0
django2-url-robots==3.1.0
docutils==0.15.2
requests==2.22.0
awscli==1.17.9
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
click==7.0
factory-boy==2.12.0
Faker==1.0.7
cryptography==2.8
idna==2.8
whitenoise==5.0.1
celery~=4.4
amqp==2.5.2
